+++ CHANGELOG.md
@@ -123,3 +123,4 @@
 * Define `Can` and `Cannot` to define `Default` instances for `Powers` and `Restrictions`
 * Introduce `Returning` type family to automatically decode the API response
 * Move `Poll` status info from `Poll` datatype to `Polling` constructor of `Content` datatype
+* Define `Persistable` instance for sending `Photo`